third_party/WebKit/LayoutTests/http/tests/devtools/tracing/timeline-paint/timeline-paint-with-layout-invalidations-on-deleted-node-expected.txt

Tests the Timeline API instrumentation of layout invalidations on a deleted node.


Running: testLocalFrame
paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:17}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "BODY"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:19}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:20}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
    {
        cause : {reason: Removed from layout, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:21}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "#text"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
    {
        cause : {reason: Removed from layout, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:21}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV id='testElement'"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]

Running: testSubframe
second paint invalidations[
    {
        cause : {reason: Inline CSS style declaration was mutated, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:28}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : ""
        nodeName : "DIV"
        selectorPart : undefined
        type : "StyleRecalcInvalidationTracking"
    }
    {
        cause : {reason: Style changed, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:29}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
    {
        cause : {reason: Removed from layout, stackTrace: test://evaluations/0/timeline-paint-with-layout-invalidations-on-deleted-node.js:30}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "LayoutInvalidationTracking"
    }
]
